move creator to new package, comment callback, start to add shorter habit holder
