+++ app/src/main/java/skills/future/planer/ui/goals/creator/GoalsCreatorActivity.java
@@ -1,4 +1,4 @@
-package skills.future.planer.ui.goals;
+package skills.future.planer.ui.goals.creator;
 
 import android.app.DatePickerDialog;
 import android.os.Bundle;
