+++ app/src/main/res/layout/activity_goals_creator.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.goals.GoalsCreatorActivity">
+    tools:context=".ui.goals.creator.GoalsCreatorActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/saveFABGoalCreator"
